Do not assume destruction of ApduService after onDeactivated was called
If HostApduService.onDeactivated is called, the current connection to
the reader ended, however, this does not mean that the HostApduService
instance is going to be destructed. If a new connection is initiated by
the reader and going to routed to the same HCE app, the existing
HostApduService instance may be reused.
Since there only ever exists one ApduService, setting the reference in
the constructor is enough.