Fix hostname verification for IP addresses
Do not try to do a reverse lookup on the IP address to get a hostname,
but use the literal IP address. This fixes TLS certificate verification
if the user entered an IP address as server address, the TLS certificate
contains this IP address, and the rDNS record is set.